+++ src/PSBar.cpp
@@ -15,9 +15,9 @@ PSBar::PSBar(singleHits *h1, singleHits *h2) {}
 // Getters
 UInt_t PSBar::GetQNear() { return 0; }
 UInt_t PSBar::GetQFar() { return 0; }
-float PSBar::GetQMean() { return 0; }
-float PSBar::GetQMeanShort() { return 0; }
-float PSBar::GetPSD() { return 0; }
+double PSBar::GetQMean() { return 0; }
+double PSBar::GetQMeanShort() { return 0; }
+double PSBar::GetPSD() { return 0; }
 Long_t PSBar::GetDelT() const { return 0; }
 Long_t PSBar::GetTStampNear() { return 0; }
 Long_t PSBar::GetTStampFar() { return 0; }
