+++ src/globals.cpp
@@ -7,6 +7,6 @@ UShort_t GateStart = 213;      // NaI 500MSPS has *2 ns
 UShort_t GateLenLong = 2000;   // NaI 500MSPS has *2 ns
 UShort_t GateLenShort = 275;   // NaI 500MSPS has *2 ns
 UShort_t PairCoincWindow = 96; // in ns
-float EvalNormFactor =
+double EvalNormFactor =
     31.4; // Factor to scale Eval energy with Digitizer energy
 } // namespace digiAnalysis
